refactor!: Adjust config struct for multi-source trait
BREAKING CHANGE: The configuration format has changed significantly and
is not backwards compatible. Please refer to the sample configuration for
an up-to-date example. Notably, the `famedly` configuration struct is now
renamed to `zitadel`, `ldap` is now nested under `sources`, `cache_path`
is now under `ldap`, and `sources` have also configs for `csv` and `ukt`.